+++ GabrovinaShell.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <Windows.h>
-#include "comms.h"
+#include "src/headers/comms.h"
 
 char version[] = "v1.0";
