--- GabrovinaShell.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.5.33424.131
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "GabrovinaShell", "GabrovinaShell.vcxproj", "{3BB9A6A1-5437-42D6-8EC0-89B0FFE7D940}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{3BB9A6A1-5437-42D6-8EC0-89B0FFE7D940}.Debug|x64.ActiveCfg = Debug|x64
-		{3BB9A6A1-5437-42D6-8EC0-89B0FFE7D940}.Debug|x64.Build.0 = Debug|x64
-		{3BB9A6A1-5437-42D6-8EC0-89B0FFE7D940}.Debug|x86.ActiveCfg = Debug|Win32
-		{3BB9A6A1-5437-42D6-8EC0-89B0FFE7D940}.Debug|x86.Build.0 = Debug|Win32
-		{3BB9A6A1-5437-42D6-8EC0-89B0FFE7D940}.Release|x64.ActiveCfg = Release|x64
-		{3BB9A6A1-5437-42D6-8EC0-89B0FFE7D940}.Release|x64.Build.0 = Release|x64
-		{3BB9A6A1-5437-42D6-8EC0-89B0FFE7D940}.Release|x86.ActiveCfg = Release|Win32
-		{3BB9A6A1-5437-42D6-8EC0-89B0FFE7D940}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {FF50EB12-8F55-4E7B-82E6-714A9261BFC1}
-	EndGlobalSection
-EndGlobal
