+++ src/c_files/comms.c
@@ -1,4 +1,4 @@
-#include "comms.h"
+#include "src/headers/comms.h"
 
 char ver2[] = "v1.0";
