+++ src/c_files/errhandler.c
@@ -1,5 +1,5 @@
 #pragma once
-#include "errhandler.h"
+#include "src/headers/errhandler.h"
 
 void getcurrdirerror(void) {
